requirements.txt
langchain_community
langchain-google-genai
langchain
langchainhub